+++ src/CMakeLists.txt
@@ -3,14 +3,13 @@
 #
 
 # Add source to this project's executable.
-set(SRC "Game.cpp"
-	"EntityRef.cpp"
-	"Events.cpp"
-	"Component.cpp"
-	"ComponentRef.cpp"
-	"RenderComponent.cpp"
-    "PrinterComponent.cpp"
-	"SquareRenderComponent.cpp" "AudioTestComponent.cpp")
+
+
+file(GLOB_RECURSE ENGINE_SRC ${CMAKE_SOURCE_DIR}/src/engine/*.cpp)
+file(GLOB_RECURSE COMPONENTS_SRC ${CMAKE_SOURCE_DIR}/src/components/*.cpp)
+file(GLOB_RECURSE PHYSICS_SRC ${CMAKE_SOURCE_DIR}/src/physics/*.cpp)
+
+set(SRC ${ENGINE_SRC} ${COMPONENTS_SRC} ${PHYSICS_SRC} "Game.cpp")
 
 
 if (WIN32)
@@ -27,9 +26,10 @@ endif()
 # Set C++ standard to 20
 target_compile_features(LightsOut PUBLIC cxx_std_20)
 
-# Link to the actual SDL3 library.
+# Link to the libraries used.
 target_link_libraries(LightsOut PRIVATE SDL3::SDL3)
 target_link_libraries(LightsOut PRIVATE SDL3_ttf::SDL3_ttf)
 target_link_libraries(LightsOut PRIVATE SDL3_mixer::SDL3_mixer)
+
 
 target_include_directories(LightsOut PRIVATE "include")
